Tg qbe backend main branch (#307)
* Step 4: added new field DeclaredClaims

* Step 6: added new fields for bond and clause

* Fix to DeclaireClaims struct

* Added new fields: YearOfEstablishment, CityCode, EnterpriseAddress, BuildingAddress

* Added new bool field IsNaicsSellable

* wip: check QBE input data

* fix to checkDeclaredClaims()

* Qbe checks moved inside CommercialCombinedProduct switch case

* proper type for policy.HasBond

* added chacks for bonds, statements and signatory

* workEmployersRemuneration, NorthAmericanMarket e totalBilled/revenue converted to float64

* More checks to be sure some pointers are not nil

* IsAlarm -> HasAlarm

---------

+++ companydata/global.go
@@ -195,7 +195,7 @@ func getDeductableMap(key string) string {
 	}
 	return res
 }
-func getMapRevenue(key int) string {
+func getMapRevenue(key float64) string {
 	var res int
 
 	if key <= 200000 { // will be false if person is not in the map
